# docs/swagger.yaml
basePath: /
definitions:
  order.Order:
    properties:
      customer_id:
        type: string
      id:
        type: string
      order_status:
        type: string
    required:
    - customer_id
    - order_status
    type: object
  user.User:
    properties:
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - firstname
    - lastname
    - password
    - username
    type: object
  user.UserSignUpDTO:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  v1.IdResponse:
    properties:
      id:
        type: string
    type: object
  v1.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for orders usecase
  title: Order service
  version: "1.0"
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: authentification account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.UserSignUpDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: SignUp
      tags:
      - auth
  /v1/order:
    delete:
      description: Delete order by ID
      parameters:
      - description: Order ID
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Delete order
      tags:
      - order
    get:
      description: Get order by ID
      parameters:
      - description: Order ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success get order
          schema:
            $ref: '#/definitions/order.Order'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Get order
      tags:
      - order
    post:
      consumes:
      - application/json
      description: Create order
      parameters:
      - description: Order object
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.Order'
      produces:
      - application/json
      responses:
        "200":
          description: Success create
          schema:
            $ref: '#/definitions/v1.IdResponse'
        "400":
          description: Invalid body
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Create order
      tags:
      - order
    put:
      consumes:
      - application/json
      description: Update order
      parameters:
      - description: Order object
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid body
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Update order
      tags:
      - order
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"